requirements.txt
flask
ibm-watsonx-ai
langchain-ibm
python-dotenv
pdfkit